+++ app/controllers/myaccount/RecoverPasswordController.java
@@ -1,0 +1,53 @@
+package controllers.myaccount;
+
+import com.commercetools.sunrise.framework.components.controllers.PageHeaderControllerComponentSupplier;
+import com.commercetools.sunrise.framework.components.controllers.RegisteredComponents;
+import com.commercetools.sunrise.framework.controllers.cache.NoCache;
+import com.commercetools.sunrise.framework.controllers.metrics.LogMetrics;
+import com.commercetools.sunrise.framework.reverserouters.myaccount.authentication.AuthenticationReverseRouter;
+import com.commercetools.sunrise.framework.template.TemplateControllerComponentsSupplier;
+import com.commercetools.sunrise.framework.template.engine.ContentRenderer;
+import com.commercetools.sunrise.myaccount.authentication.recoverpassword.recover.RecoverPasswordControllerAction;
+import com.commercetools.sunrise.myaccount.authentication.recoverpassword.recover.RecoverPasswordControllerComponent;
+import com.commercetools.sunrise.myaccount.authentication.recoverpassword.recover.RecoverPasswordFormData;
+import com.commercetools.sunrise.myaccount.authentication.recoverpassword.recover.SunriseRecoverPasswordController;
+import com.commercetools.sunrise.myaccount.authentication.recoverpassword.recover.viewmodels.RecoverPasswordPageContentFactory;
+import io.sphere.sdk.customers.CustomerToken;
+import play.data.FormFactory;
+import play.mvc.Result;
+
+import javax.annotation.Nullable;
+import javax.inject.Inject;
+import java.util.concurrent.CompletionStage;
+
+@LogMetrics
+@NoCache
+@RegisteredComponents({
+        TemplateControllerComponentsSupplier.class,
+        PageHeaderControllerComponentSupplier.class,
+        RecoverPasswordControllerComponent.class
+})
+public final class RecoverPasswordController extends SunriseRecoverPasswordController {
+    private final AuthenticationReverseRouter authenticationReverseRouter;
+
+    @Inject
+    public RecoverPasswordController(final ContentRenderer contentRenderer, final FormFactory formFactory,
+                                     final RecoverPasswordPageContentFactory pageContentFactory,
+                                     final RecoverPasswordFormData formData,
+                                     final RecoverPasswordControllerAction controllerAction,
+                                     final AuthenticationReverseRouter authenticationReverseRouter) {
+        super(contentRenderer, formFactory, pageContentFactory, formData, controllerAction);
+        this.authenticationReverseRouter = authenticationReverseRouter;
+    }
+
+    @Nullable
+    @Override
+    public String getTemplateName() {
+        return "my-account-forgot-password";
+    }
+
+    @Override
+    public CompletionStage<Result> handleSuccessfulAction(final CustomerToken output, final RecoverPasswordFormData formData) {
+        return redirectToCall(authenticationReverseRouter.logInProcessCall());
+    }
+}
