+++ conf/routes
@@ -97,9 +97,9 @@ GET         /:languageTag/user/logout                                           
 # ~~~~~
 
 # Show form to enter password recover email
-GET         /:languageTag/password/recovery                                     @controllers.myaccount.PasswordRecoveryController.show(languageTag: String)
+GET         /:languageTag/password/recovery                                     @controllers.myaccount.RecoverPasswordController.show(languageTag: String)
 # Send recovery password email
-POST         /:languageTag/password/recovery                                    @controllers.myaccount.PasswordRecoveryController.process(languageTag: String)
+POST         /:languageTag/password/recovery                                    @controllers.myaccount.RecoverPasswordController.process(languageTag: String)
 
 # View password reset
 GET         /:languageTag/password/reset/:token                                 @controllers.myaccount.ResetPasswordController.show(languageTag: String, token: String)
