+++ my-account/app/com/commercetools/sunrise/myaccount/authentication/recoverpassword/recover/DefaultRecoverPasswordFormData.java
@@ -1,0 +1,24 @@
+package com.commercetools.sunrise.myaccount.authentication.recoverpassword.recover;
+
+import play.data.validation.Constraints.Email;
+
+/**
+ * Default implementation of {@link RecoverPasswordFormData}
+ */
+public class DefaultRecoverPasswordFormData implements RecoverPasswordFormData {
+    @Email
+    private String email;
+
+    @Override
+    public String email() {
+        return email;
+    }
+
+    public String getEmail() {
+        return email;
+    }
+
+    public void setEmail(final String email) {
+        this.email = email;
+    }
+}
